llm-api/qwen/qwen-long/document_understanding/file_upload_to_dashscope.py
import os
from pathlib import Path

from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()
client = OpenAI(
    api_key=os.getenv("DASHSCOPE_API_KEY"),  # 如果您没有配置环境变量，请在此处替换您的API-KEY
    base_url="https://dashscope.aliyuncs.com/compatible-mode/v1",  # 填写DashScope服务base_url
)

file_object = client.files.create(file=Path("中央和国家机关差旅住宿费和伙食补助费标准表.pdf"), purpose="file-extract")
print(file_object.id)
