llm-api/pyproject.toml
[project]
name = "llm-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
]
